TargetMarketingTools-backend/Tmarket/requirements.txt
asgiref~=3.2.10
backports.zoneinfo==0.2.1
Django==3.1.2
django-cors-headers~=3.10.1
djangorestframework~=3.13.1
psycopg2-binary~=2.9.3
PyJWT~=2.3.0
pytz~=2022.1
sqlparse~=0.4.2
requests~=2.27.1
